add actionable steps

+++ app/Http/Controllers/IdeaController.php
@@ -45,7 +45,11 @@ class IdeaController extends Controller
      */
     public function store(StoreIdeaRequest $request)
     {
-        Auth::user()->ideas()->create($request->validated());
+        $idea = Auth::user()->ideas()->create($request->safe()->except('steps'));
+
+        $idea->steps()->createMany(
+            collect($request->steps)->map(fn ($step) => ['description' => $step])
+        );
 
         return to_route('idea.index')->with('success', 'Idea Created!');
 
@@ -82,7 +86,8 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
-        // auth that it's allowed
+        $this->authorize('delete', $idea);
+
         $idea->delete();
 
         return to_route('idea.index');
